requirements.txt
beautifulsoup4==4.7.1
django-cors-headers==2.2.0
djangorestframework==3.7.7
django-static-precompiler==1.8.2
requests==2.9.1
six==1.10.0
ssh-import-id==5.5
subdomains==2.1.1
django-debug-toolbar==1.9.1
PyYAML==3.12
backcall==0.1.0
chardet==2.3.0
decorator==4.3.2
Django==2.1.7
ipython==7.3.0
ipython-genutils==0.2.0
jedi==0.13.3
parso==0.3.4
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
psycopg2-binary==2.7.7
ptyprocess==0.6.0
Pygments==2.3.1
pytz==2018.9
traitlets==4.3.2
urllib3==1.13.1
wcwidth==0.1.7
mock==2.0.0
git+git://github.com/michdol/django-subdomains.git#egg=django-subdomains
